move loader to hop
