+++ src/aioway/_media/audios.py
@@ -10,8 +10,8 @@ from torch.utils import data
 from torchcodec import decoders as dec
 
 from aioway._utils import current_fake_mode, torch_set_fake_mode_func
-from aioway.spaces import Attr
 from aioway.dsets import IsStftTag, SampleRateTag
+from aioway.spaces import Attr
 
 from ._av import AudioStream
 from ._bases import TorchCompatible
