+++ src/aioway/_media/videos.py
@@ -9,8 +9,8 @@ import torch
 from torchcodec import decoders as dec
 
 from aioway._utils import current_fake_mode, num_threads, torch_set_fake_mode_func
-from aioway.spaces import Attr
 from aioway.dsets import IsVideoTag
+from aioway.spaces import Attr
 
 from ._av import VideoStream
 from ._bases import TorchCompatible
