+++ src/aioway/dsets/dsets.py
@@ -11,16 +11,11 @@ import tensordict as td
 import torch
 from torch.utils import data
 
-from aioway._utils import HasLen
-from aioway.hop import Hop, TdictHop, TensorHop, hop_dcls
+from aioway.hop import LoaderHop, LoaderOpt, TdictLoaderHop, TensorLoaderHop
 
 from .tags import TagDict
 
 __all__ = [
-    "LoaderOpt",
-    "LoaderHop",
-    "TensorLoaderHop",
-    "TdictLoaderHop",
     "Dset",
     "Stream",
     "Frame",
@@ -30,84 +25,6 @@ __all__ = [
     "TensorFrame",
     "TdictFrame",
 ]
-
-
-@dcls.dataclass(frozen=True)
-class LoaderOpt:
-    """
-    The optoins for `data.DataLoader` on `Frame` in `FrameStream`.
-    """
-
-    batch_size: int = 1
-    "The batch size of individual batches."
-
-    drop_last: bool = False
-    "Whether to drop the last batch, which may have a different `batch_size`."
-
-    shuffle: bool = False
-    "To shuffle or not."
-
-    sampler: data.Sampler[int] | None = None
-    "How to sample in case when want to shuffle."
-
-
-@hop_dcls
-class LoaderHop[T = typing.Any](Hop[T]):
-    """
-    A `Hop` backed by a `torch` `DataLoader`.
-    """
-
-    _: dcls.KW_ONLY
-
-    dset: data.Dataset[T]
-    """
-    The data loader that would be iterated over.
-    """
-
-    opts: LoaderOpt = LoaderOpt()
-    """
-    The options to pass to `DataLoader`.
-    """
-
-    @property
-    @typing.override
-    def size(self) -> int:
-        # May be a frame or stream, so we need to test.
-        # Perhaps this is a bad design.
-        if not isinstance(self.dset, HasLen):
-            return NotImplemented
-
-        # Drop last or not would affect counts.
-        count, remain = divmod(len(self.dset), self.opts.batch_size)
-
-        if remain and not self.opts.drop_last:
-            count += 1
-
-        return count
-
-    def iterate(self) -> cabc.Generator[T]:
-        yield from self._dataloader()
-
-    def _dataloader(self) -> data.DataLoader[T]:
-        return data.DataLoader(self.dset, **dcls.asdict(self.opts))
-
-
-@hop_dcls
-class TensorLoaderHop(LoaderHop[torch.Tensor], TensorHop):
-    """
-    A `Hop` to load `torch.Tensor`.
-    """
-
-
-@hop_dcls
-class TdictLoaderHop(LoaderHop[td.TensorDict], TdictHop):
-    """
-    A `Hop` to load `td.TensorDict`.
-    """
-
-    @typing.override
-    def _dataloader(self) -> data.DataLoader[td.TensorDict]:
-        return data.DataLoader(self.dset, collate_fn=td.stack, **dcls.asdict(self.opts))
 
 
 @typing.dataclass_transform(frozen_default=True)
