+++ src/aioway/hop/__init__.py
@@ -3,5 +3,6 @@
 from .funcs import *
 from .hop import *
 from .iters import *
+from .loaders import *
 from .nn import *
 from .relalg import *
