+++ tests/compilers/test_emitters.py
@@ -6,11 +6,11 @@ import pytest
 from torch import nn
 
 from aioway.algos import SupervisedAlgo
-from aioway.spaces import Attr
 from aioway.compilers import JustLinearEmitter
 from aioway.dsets import AttrTag, Stream, TagDict, TensorListHop
 from aioway.hop import Linear, ListHop, NnHop, NnInit, NnLayerHop, TensorHop
 from aioway.sinks import Sink
+from aioway.spaces import Attr
 
 
 @pytest.fixture
